Use existing package name in test data config
`npm show` failed with error `'@keep-network/test-package' is not in the
npm registry.`.
Here we update the package in test data package.json to use actual
package published to the registry.

We may consider supporting this case in the future, but it's not
neccessary now as we are bumping versions for already published
packages.