Filter versions to match prerelease ID
In a scenario where there are different prerelease versions with
multiple prerelease ID s published we want to be able to increment them
independently. So for example if we're trying to publish version
1.8.0-pre and there are already published:
1.8.0-pre.0, 1.8.0-pre.1, 1.8.0-rc.0
We want the `pre` version to be incemented to `1.8.0-pre.2`.
The current solution didn't allow it to happen as it returned `rc`
version as the newest.